BI-6701 Renaming retry package to producer
