+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/producer/AvroSerializer.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.chipsrestinterfacesconsumer.retry;
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.producer;
 
 import org.apache.avro.io.DatumWriter;
 import org.apache.avro.io.Encoder;
